requirements.txt
playwright
Pillow
python-dotenv
aiohttp
aiohttp-socks